requirement.txt
asgiref==3.4.1
astroid==2.6.6
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-appconf==1.0.4
django-cors-headers==3.7.0
django-imagekit==4.0.2
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pilkit==2.0
Pillow==8.3.1
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.9.6
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
